Adjusted maxOptimizationDistance in app.vue to compute the max. possible steps that can be made given an unoptimized config while also considering numeric features.
Renamed selectedOptimizationConfigName from uiLogicMixin.js to selectedUnoptimizedConfigName.

Made selectedUnoptimizedConfig in uiLogicMixin.js computed prop instead of data so that the unoptimized config and therefore and more importantly the maxOptimizationDistance in app.vue reacts to changes of the selection of a config to be optimized (the dropdown for config selection in the modal).

Added reset for the selectedUnoptimizedConfig in uiLogicMixin.js in closeOptimizationModal to empty string to prevent unwanted preselection the next time the OptimizationModal.vue is opened.