Replaced Collapse-Button implementation via primevue with implementation via bootstrap vue and fontawesome icons.
Moved lines for config card in ConfigArea.vue to ConfigCard.vue and adjusted code of ConfigCard.vue to be more versatile. the component ConfigCard.vue can now not only be used for dialog previews (showing alternatives or optimizations) but also in the ConfigArea.vue.

Added ConfigCard.vue component with props to ConfigArea.vue.

Chip.vue: removed this.visible = false, because it messed up the visible chips/features that are rendered in the config card ("fake double remove")